Added Fancy Blood Hud, Added New Skill, Added Charm and Slot

+++ src/main/java/com/huto/hemomancy/gui/GuiUtil.java
@@ -12,7 +12,7 @@ public class GuiUtil {
 	/**
 	 * Draws a textured rectangle at the current z-value. Ported From past Versions
 	 */
-	public static void drawTexturedModalRect(int x, int y, int textureX, int textureY, int width, int height) {
+	public static void drawTexturedModalRect(float x, float y, float textureX, float textureY, float width, float height) {
 		/*
 		 * float f = 0.00390625F; float f1 = 0.00390625F;
 		 */
